Added code to load a pattern from a file to GameUtils
GameUtils.loadPattern(File) accepts a File object which should point
to a CSV file compliant with the specification in
`docs/file_format.md`. The file is converted into a board of organisms
which the method returns. The board is the smallest possible to fit
all the organisms in the file.

Other changes:

- Updated the file format specification in `docs/file_format.md` to be
  more specific
- Added a simple test pattern in `data/patterns/test.csv`
- Renamed `Organisms` to `organism` (the package and the directory) to
  match with Java standard practices