Bugfix movement - start with empty board and copy over as necessary.
The current code copied the current state of the board to create the
next generation. This behaviour was incorrect, as it is impossible to
distinguish between the old space of the organism and a new organism
that has moved into the new space

For example:
Original
+---+---+
| A | B |
+---+---+
|   |   |
+---+---+

A moves to the spot currently occupied by B, and B moves to one of the
empty spaces.

Expected result:
+---+---+
|   | A |
+---+---+
|   | B |
+---+---+

However, under the old algorithm, when B moves to its new space, the
old one is emptied regardless of wether a new organism moved into that
space.

Actual result:
+---+---+
|   |   |
+---+---+
|   | B |
+---+---+

Also:
- Sanity checks on the wrapped x and y coordinates
- Simplified wrapping code in the non-negative case
- Documented the move method in Fish.java